% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyCars.R
\docType{data}
\name{tidy_cars}
\alias{tidy_cars}
\title{Descriptive data title}
\format{a \code{data.frame} containing the following fields:
\describe{
\item{speed_mph}{}
\item{stopping_distace}{}
}}
\source{
The data comes from________________________.
}
\description{
the stopping distances of cars at different speeds.
}
\seealso{
\link{TidyCars}
}
