% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyCars.R
\docType{data}
\name{tidy_iris}
\alias{tidy_iris}
\title{Descriptive data title}
\format{a \code{data.frame} containing the following fields:
\describe{
\item{sepal_length}{}
\item{Sepal.Width}{}
\item{Petal.Length}{}
\item{Petal.Width}{}
\item{species}{}
}}
\source{
The data comes from________________________.
}
\description{
new data file in this packages
}
\seealso{
\link{TidyCars}
}
