% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyCars.R
\docType{package}
\name{TidyCars}
\alias{TidyCars}
\alias{TidyCars-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
TidyCars
A data package for TidyCars.

Use \code{data(package='TidyCars')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{tidy_iris}
}
