% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyCars.R
\docType{data}
\name{tidy_cars}
\alias{tidy_cars}
\title{The stopping distances of cars traveling at different speeds.}
\format{a \code{data.frame} containing the following fields:
\describe{
\item{speed_mph}{The speed of the vehicle.}
\item{stopping_distace}{The stopping distance of the vehicle.}
}}
\source{
The data comes from the cars data set distributed with R.
}
\description{
The stopping distances of cars at different speeds.
}
\seealso{
\link{TidyCars}
}
