% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidyCars.R
\docType{package}
\name{TidyCars}
\alias{TidyCars}
\alias{TidyCars-package}
\title{A tidied cars data set.}
\description{
Cars but better. The variable names are more meaninful.
}
\details{
TidyCars
A data package for TidyCars.

The columns have been renamed to indicate the units and better describe what is measured.
}
\seealso{
\link{tidy_cars}
}
